developed key backend functionalities for payment,Developed  API endpoints

+++ ecommerce/src/main/java/com/xag/model/Order.java
@@ -41,7 +41,7 @@ public class Order {
 
     private double totalMrpPrice;
 
-    private double totalSellingPrice;
+    private Integer totalSellingPrice;
 
     private Integer discount;
